ADD--events of aggregate order

+++ src/main/java/com/sofkau/ddd/order/values/Payment.java
@@ -3,9 +3,9 @@ package com.sofkau.ddd.order.values;
 import co.com.sofka.domain.generic.ValueObject;
 import com.sofkau.ddd.drink_detail.values.Value;
 
-public class Payment implements ValueObject<Integer> {
+public class Payment implements ValueObject<Double> {
     @Override
-    public Integer value() {
+    public Double value() {
         return null;
     }
 }
